Fix #45, Move cmds and utils into separate files
