+++ CMakeLists.txt
@@ -1,7 +1,9 @@
 project(CFE_SAMPLE_APP C)
 
 # Create the app module
-add_cfe_app(sample_app fsw/src/sample_app.c)
+add_cfe_app(sample_app fsw/src/sample_app.c
+                       fsw/src/sample_app_cmds.c
+                       fsw/src/sample_app_utils.c)
 
 # Include the public API from sample_lib to demonstrate how
 # to call library-provided functions
