+++ fsw/src/sample_app.h
@@ -88,6 +88,11 @@ typedef struct
     CFE_TBL_Handle_t TblHandles[SAMPLE_APP_NUMBER_OF_TABLES];
 } SAMPLE_APP_Data_t;
 
+/*
+** Global data structure
+*/
+extern SAMPLE_APP_Data_t SAMPLE_APP_Data;
+
 /****************************************************************************/
 /*
 ** Local function prototypes.
@@ -100,13 +105,5 @@ int32 SAMPLE_APP_Init(void);
 void  SAMPLE_APP_ProcessCommandPacket(CFE_SB_Buffer_t *SBBufPtr);
 void  SAMPLE_APP_ProcessGroundCommand(CFE_SB_Buffer_t *SBBufPtr);
 int32 SAMPLE_APP_ReportHousekeeping(const CFE_MSG_CommandHeader_t *Msg);
-int32 SAMPLE_APP_ResetCounters(const SAMPLE_APP_ResetCountersCmd_t *Msg);
-int32 SAMPLE_APP_Process(const SAMPLE_APP_ProcessCmd_t *Msg);
-int32 SAMPLE_APP_Noop(const SAMPLE_APP_NoopCmd_t *Msg);
-void  SAMPLE_APP_GetCrc(const char *TableName);
-
-int32 SAMPLE_APP_TblValidationFunc(void *TblData);
-
-bool SAMPLE_APP_VerifyCmdLength(CFE_MSG_Message_t *MsgPtr, size_t ExpectedLength);
 
 #endif /* SAMPLE_APP_H */
