+++ fsw/src/sample_app_utils.h
@@ -18,41 +18,19 @@
 
 /**
  * @file
- *
- * Common definitions for all sample_app coverage tests
+ *   This file contains the prototypes for the Sample App utility functions
  */
 
-#ifndef SAMPLE_APP_COVERAGETEST_COMMON_H
-#define SAMPLE_APP_COVERAGETEST_COMMON_H
+#ifndef SAMPLE_APP_UTILS_H
+#define SAMPLE_APP_UTILS_H
 
 /*
- * Includes
- */
-
-#include "utassert.h"
-#include "uttest.h"
-#include "utstubs.h"
-
-#include "cfe.h"
+** Required header files.
+*/
 #include "sample_app.h"
-#include "sample_app_cmds.h"
-#include "sample_app_utils.h"
-#include "sample_app_events.h"
-#include "sample_app_table.h"
 
-/*
- * Macro to add a test case to the list of tests to execute
- */
-#define ADD_TEST(test) UtTest_Add((Test_##test), Sample_UT_Setup, Sample_UT_TearDown, #test)
+bool  SAMPLE_APP_VerifyCmdLength(CFE_MSG_Message_t *MsgPtr, size_t ExpectedLength);
+int32 SAMPLE_APP_TblValidationFunc(void *TblData);
+void  SAMPLE_APP_GetCrc(const char *TableName);
 
-/*
- * Setup function prior to every test
- */
-void Sample_UT_Setup(void);
-
-/*
- * Teardown function after every test
- */
-void Sample_UT_TearDown(void);
-
-#endif /* SAMPLE_APP_COVERAGETEST_COMMON_H */
+#endif /* SAMPLE_APP_UTILS_H */
