+++ unit-test/coveragetest/coveragetest_sample_app.c
@@ -35,7 +35,6 @@
 /*
  * Includes
  */
-
 #include "sample_lib.h" /* For SAMPLE_LIB_Function */
 #include "sample_app_coveragetest_common.h"
 #include "ut_sample_app.h"
